refactor: move shared layout types from lib/Epub to lib/Typesetter
