+++ lib/Typesetter/Typesetter/FootnoteEntry.h
