+++ lib/Typesetter/Typesetter/blocks/Block.h
@@ -4,7 +4,7 @@ class GfxRenderer;
 
 typedef enum { TEXT_BLOCK, IMAGE_BLOCK } BlockType;
 
-// a block of content in the html - either a paragraph or an image
+// a block of content - either a paragraph or an image
 class Block {
  public:
   virtual ~Block() = default;
